Therapist profile screen and its slice was created and the code was adjusted for the screen slice and navigations  in base.tsx and store.tsx
