+++ navigations-maps/Base.tsx
@@ -4,14 +4,14 @@ import SignIn from "../screens/signin-screen/SignIn";
 import SignUp from "../screens/signup-screen/SignUp";
 import LogoutScreen from "../screens/LogoutScreen/logout";
 import DrawerNavigator from "../navigators/DrawerNavigator";
-import TherapistProfile from "../screens/therapistProfileScreen/HomeScreen" // Import your DrawerNavigator
+import TherapistScreen from "../screens/therapistProfileScreen/TherapistScreen" // Import your DrawerNavigator
 
 export const BaseRouteNames = {
   SignIn: "SignIn",
   SignUp: "SignUp",
   Logout: "Logout",
   Home: "Home",
-  TherapistProfile:"TherapistProfile" // This will now be the DrawerNavigator
+  TherapistProfile:"TherapistScreen" // This will now be the DrawerNavigator
 } as const;
 
 export type BaseRouteName = typeof BaseRouteNames[keyof typeof BaseRouteNames];
@@ -52,7 +52,7 @@ export const BaseRoutes: IRoute[] = [
     },
   },
   {
-    component: TherapistProfile, // Changed from Therapist to DrawerNavigator
+    component: TherapistScreen, // Changed from Therapist to DrawerNavigator
     title: BaseRouteNames.TherapistProfile,
     options: {
       headerShown: true, // Hide header since DrawerNavigator manages its own headers
