+++ store/store.ts
@@ -4,6 +4,7 @@ import homeScreenReducer from "../screens/HomeScreen/homeScreenSlice";
 import { signInSlice } from "../screens/signin-screen/SignInSlice";
 import { signUpSlice } from "../screens/signup-screen/SignUpSlice";
 import { logoutSlice } from "../screens/LogoutScreen/LogoutSlice";
+import { therapistSlice } from "../screens/therapistProfileScreen/therapistSlice";
 
 export const store = configureStore({
   reducer: {
@@ -12,6 +13,7 @@ export const store = configureStore({
     signUp: signUpSlice.reducer,
     logout: logoutSlice.reducer,
     homeScreen: homeScreenReducer,
+    therapist: therapistSlice.reducer,
   },
 });
